moved game2 components into separate folder
