+++ src/components/styles/Game.scss
@@ -19,7 +19,10 @@
 
   margin: 10px;
 
-  padding: 5px;
+  padding-left: 50px;
+  padding-right: 50px;
+  padding-top: 20px;
+  padding-bottom: 50px;
 
   border-style: solid;
   border-radius: 30px;
